refactor: Rename validator package to validators
