+++ src/mg/emberframework/util/validation/ValidatorRegistry.java
@@ -9,13 +9,13 @@ import mg.emberframework.annotation.validation.Email;
 import mg.emberframework.annotation.validation.Length;
 import mg.emberframework.annotation.validation.Numeric;
 import mg.emberframework.annotation.validation.Required;
-import mg.emberframework.util.validation.validator.DateValidator;
-import mg.emberframework.util.validation.validator.EmailValidator;
-import mg.emberframework.util.validation.validator.FieldValidator;
-import mg.emberframework.util.validation.validator.LengthValidator;
-import mg.emberframework.util.validation.validator.NeutralValidator;
-import mg.emberframework.util.validation.validator.NumericValidator;
-import mg.emberframework.util.validation.validator.RequiredValidator;
+import mg.emberframework.util.validation.validators.DateValidator;
+import mg.emberframework.util.validation.validators.EmailValidator;
+import mg.emberframework.util.validation.validators.FieldValidator;
+import mg.emberframework.util.validation.validators.LengthValidator;
+import mg.emberframework.util.validation.validators.NeutralValidator;
+import mg.emberframework.util.validation.validators.NumericValidator;
+import mg.emberframework.util.validation.validators.RequiredValidator;
 
 public class ValidatorRegistry {
     private ValidatorRegistry() {
